fixed forge build

+++ packages/contracts/src/dollar/DirectGovernanceFarmer.sol
@@ -19,9 +19,9 @@ import "./interfaces/IUbiquityAlgorithmicDollarManager.sol";
 contract DirectGovernanceFarmer is ReentrancyGuard {
     using SafeERC20 for IERC20;
 
-    address public token2;//USDT decimal 6
-    address public token1;//USDC decimal 6
-    address public token0;//DAI
+    address public token2; //USDT decimal 6
+    address public token1; //USDC decimal 6
+    address public token0; //DAI
     address public ubiquity3PoolLP;
     address public ubiquityDollar;
     address public depositZapUbiquityDollar;
@@ -42,7 +42,6 @@ contract DirectGovernanceFarmer is ReentrancyGuard {
         address token,
         uint256 amount
     );
-
 
     constructor(address _manager, address base3Pool, address depositZap) {
         manager = IUbiquityAlgorithmicDollarManager(_manager); // 0x4DA97a8b831C345dBe6d16FF7432DF2b7b776d98
@@ -75,7 +74,6 @@ contract DirectGovernanceFarmer is ReentrancyGuard {
         return this.onERC1155Received.selector;
     }
 
-
     /**
      * @dev Deposit into Ubiquity protocol
      * @notice Stable coin (DAI / USDC / USDT / uAD) => uAD3CRV-f => Ubiquity BondingShare
@@ -85,26 +83,30 @@ contract DirectGovernanceFarmer is ReentrancyGuard {
      * @param amount Amount of tokens to deposit (For max: `uint256(-1)`)
      * @param durationWeeks Duration in weeks tokens will be locked (1-208)
      */
-    function deposit(
-        address token,
-        uint256 amount,
-        uint256 durationWeeks
-    ) external nonReentrant returns (uint256 stakingShareId) {
-
-
+    function deposit(address token, uint256 amount, uint256 durationWeeks)
+        external
+        nonReentrant
+        returns (uint256 stakingShareId)
+    {
         // DAI / USDC / USDT / UAD
-        require(isMetaPoolCoin(token), "Invalid token: must be DAI, USD Coin, Tether, or Ubiquity Dollar");
+        require(
+            isMetaPoolCoin(token),
+            "Invalid token: must be DAI, USD Coin, Tether, or Ubiquity Dollar"
+        );
         require(amount > 0, "amount must be positive vale");
-        require(durationWeeks >= 1 && durationWeeks <= 208, "duration weeks must be between 1 and 208");
+        require(
+            durationWeeks >= 1 && durationWeeks <= 208,
+            "duration weeks must be between 1 and 208"
+        );
 
         //Note, due to USDT implementation, normal transferFrom does not work and have an error of "function returned an unexpected amount of data"
         //require(IERC20(token).transferFrom(msg.sender, address(this), amount), "sender cannot transfer specified fund");
         IERC20(token).safeTransferFrom(msg.sender, address(this), amount);
-        
+
         address staking = manager.bondingContractAddress();
         address stakingShare = manager.bondingShareAddress();
 
-        uint256 lpAmount;//UAD3CRVf
+        uint256 lpAmount; //UAD3CRVf
         //[UAD, DAI, USDC, USDT]
         uint256[4] memory tokenAmounts = [
             token == ubiquityDollar ? amount : 0,
@@ -112,20 +114,23 @@ contract DirectGovernanceFarmer is ReentrancyGuard {
             token == token1 ? amount : 0,
             token == token2 ? amount : 0
         ];
-        
+
         //STEP1: add DAI, USDC, USDT or uAD into metapool liquidity and get UAD3CRVf
         IERC20(token).safeIncreaseAllowance(depositZapUbiquityDollar, amount);
-        lpAmount = IDepositZap(depositZapUbiquityDollar).add_liquidity(ubiquity3PoolLP, tokenAmounts, 0);
+        lpAmount = IDepositZap(depositZapUbiquityDollar).add_liquidity(
+            ubiquity3PoolLP, tokenAmounts, 0
+        );
 
         //STEP2: stake UAD3CRVf to BondingV2
         //TODO approve token to be transferred to Bonding V2 contract
         IERC20(ubiquity3PoolLP).safeIncreaseAllowance(staking, lpAmount);
         stakingShareId = IBondingV2(staking).deposit(lpAmount, durationWeeks);
 
-        IBondingShareV2(stakingShare).safeTransferFrom(address(this), msg.sender, stakingShareId, 1, '0x');
+        IBondingShareV2(stakingShare).safeTransferFrom(
+            address(this), msg.sender, stakingShareId, 1, "0x"
+        );
 
         emit Deposit(msg.sender, token, amount, durationWeeks, stakingShareId);
-        
     }
 
     /**
@@ -136,60 +141,82 @@ contract DirectGovernanceFarmer is ReentrancyGuard {
      * @param stakingShareId Bonding Share Id to withdraw
      * @param token Token to withdraw to : DAI, USDC, USDT, 3CRV or uAD
      */
-    function withdraw(
-        uint256 stakingShareId,
-        address token
-    ) external nonReentrant returns (uint256 tokenAmount) {
-        
+    function withdraw(uint256 stakingShareId, address token)
+        external
+        nonReentrant
+        returns (uint256 tokenAmount)
+    {
         // DAI / USDC / USDT / UAD
-        require(isMetaPoolCoin(token), "Invalid token: must be DAI, USD Coin, Tether, or Ubiquity Dollar");
+        require(
+            isMetaPoolCoin(token),
+            "Invalid token: must be DAI, USD Coin, Tether, or Ubiquity Dollar"
+        );
         address staking = manager.bondingContractAddress();
         address stakingShare = manager.bondingShareAddress();
 
-        uint256[] memory stakingShareIds = IBondingShareV2(stakingShare).holderTokens(msg.sender);
+        uint256[] memory stakingShareIds =
+            IBondingShareV2(stakingShare).holderTokens(msg.sender);
         //Need to verify msg.sender by holderToken history.
         //bond.minter is this contract address so that cannot use it for verification.
-        require(isIdIncluded(stakingShareIds, stakingShareId), "sender is not true bond owner");
-        
+        require(
+            isIdIncluded(stakingShareIds, stakingShareId),
+            "sender is not true bond owner"
+        );
+
         //transfer bondingShare NFT token from msg.sender to this address
-        IBondingShareV2(stakingShare).safeTransferFrom(msg.sender, address(this), stakingShareId, 1, '0x');
-        
+        IBondingShareV2(stakingShare).safeTransferFrom(
+            msg.sender, address(this), stakingShareId, 1, "0x"
+        );
+
         // Get Bond
-        IBondingShareV2.Bond memory bond = IBondingShareV2(stakingShare).getBond(stakingShareId);
-        
+        IBondingShareV2.Bond memory bond =
+            IBondingShareV2(stakingShare).getBond(stakingShareId);
+
         // STEP 1 : Withdraw Ubiquity Bonding Shares to get back uAD3CRV-f LPs
         //address bonding = ubiquityManager.bondingContractAddress();
         IBondingShareV2(stakingShare).setApprovalForAll(staking, true);
         IBondingV2(staking).removeLiquidity(bond.lpAmount, stakingShareId);
         IBondingShareV2(stakingShare).setApprovalForAll(staking, false);
-        
+
         uint256 lpTokenAmount = IERC20(ubiquity3PoolLP).balanceOf(address(this));
-        uint256 governanceTokenAmount = IERC20(manager.governanceTokenAddress()).balanceOf(address(this));
-        
+        uint256 governanceTokenAmount =
+            IERC20(manager.governanceTokenAddress()).balanceOf(address(this));
 
         // STEP2 : Withdraw  3Crv LPs from meta pool to get back UAD, DAI, USDC or USDT
-        uint128 tokenIndex = token == ubiquityDollar ? 0 : (token == token0 ? 1 : (token == token1 ? 2 : 3));
+        uint128 tokenIndex = token == ubiquityDollar
+            ? 0
+            : (token == token0 ? 1 : (token == token1 ? 2 : 3));
         IERC20(ubiquity3PoolLP).approve(depositZapUbiquityDollar, lpTokenAmount);
-        tokenAmount = IDepositZap(depositZapUbiquityDollar).remove_liquidity_one_coin(ubiquity3PoolLP, lpTokenAmount, int128(tokenIndex), 0); //[UAD, DAI, USDC, USDT]
-        
+        tokenAmount = IDepositZap(depositZapUbiquityDollar)
+            .remove_liquidity_one_coin(
+            ubiquity3PoolLP, lpTokenAmount, int128(tokenIndex), 0
+        ); //[UAD, DAI, USDC, USDT]
+
         IERC20(token).safeTransfer(msg.sender, tokenAmount);
-        IERC20(manager.governanceTokenAddress()).safeTransfer(msg.sender, governanceTokenAmount);
+        IERC20(manager.governanceTokenAddress()).safeTransfer(
+            msg.sender, governanceTokenAmount
+        );
 
         emit Withdraw(msg.sender, stakingShareId, token, tokenAmount);
     }
 
-    function isIdIncluded(uint256[] memory idList, uint256 id) internal pure returns (bool){
-        for (uint i=0; i < idList.length; i++){
-            if (idList[i]==id){
+    function isIdIncluded(uint256[] memory idList, uint256 id)
+        internal
+        pure
+        returns (bool)
+    {
+        for (uint256 i = 0; i < idList.length; i++) {
+            if (idList[i] == id) {
                 return true;
             }
         }
         return false;
-
     }
 
     function isMetaPoolCoin(address token) public view returns (bool) {
-        return (token == token2 || token == token1 || token == token0 || token == ubiquityDollar);
+        return (
+            token == token2 || token == token1 || token == token0
+                || token == ubiquityDollar
+        );
     }
-
 }
