+++ packages/contracts/test/dollar/ERC20Ubiquity.t.sol
@@ -11,7 +11,9 @@ contract ERC20UbiquityHarness is ERC20Ubiquity {
         ERC20Ubiquity(_manager, name_, symbol_)
     {}
 
-    function exposed_transfer(address sender, address recipient, uint256 amount) external {
+    function exposed_transfer(address sender, address recipient, uint256 amount)
+        external
+    {
         _transfer(sender, recipient, amount);
     }
 }
@@ -69,8 +71,8 @@ contract ERC20UbiquityTest is LocalTestHelper {
 
     function testPermit_ShouldRevert_IfDeadlineExpired() public {
         // create owner and spender addresses
-        uint ownerPrivateKey = 0x1;
-        uint spenderPrivateKey = 0x2;
+        uint256 ownerPrivateKey = 0x1;
+        uint256 spenderPrivateKey = 0x2;
         address owner = vm.addr(ownerPrivateKey);
         address spender = vm.addr(spenderPrivateKey);
         // create owner's signature
@@ -94,21 +96,13 @@ contract ERC20UbiquityTest is LocalTestHelper {
         // run permit
         vm.prank(spender);
         vm.expectRevert("Dollar: EXPIRED");
-        ERC20Ubiquity(token_addr).permit(
-            owner,
-            spender,
-            1e18,
-            0,
-            v,
-            r,
-            s
-        );
+        ERC20Ubiquity(token_addr).permit(owner, spender, 1e18, 0, v, r, s);
     }
 
     function testPermit_ShouldRevert_IfSignatureIsInvalid() public {
         // create owner and spender addresses
-        uint ownerPrivateKey = 0x1;
-        uint spenderPrivateKey = 0x2;
+        uint256 ownerPrivateKey = 0x1;
+        uint256 spenderPrivateKey = 0x2;
         address owner = vm.addr(ownerPrivateKey);
         address spender = vm.addr(spenderPrivateKey);
         // create owner's signature
@@ -133,20 +127,14 @@ contract ERC20UbiquityTest is LocalTestHelper {
         vm.prank(spender);
         vm.expectRevert("Dollar: INVALID_SIGNATURE");
         ERC20Ubiquity(token_addr).permit(
-            owner,
-            spender,
-            1e18,
-            block.timestamp + 1 days,
-            v,
-            r,
-            s
+            owner, spender, 1e18, block.timestamp + 1 days, v, r, s
         );
     }
 
     function testPermit_ShouldIncreaseSpenderAllowance() public {
         // create owner and spender addresses
-        uint ownerPrivateKey = 0x1;
-        uint spenderPrivateKey = 0x2;
+        uint256 ownerPrivateKey = 0x1;
+        uint256 spenderPrivateKey = 0x2;
         address owner = vm.addr(ownerPrivateKey);
         address spender = vm.addr(spenderPrivateKey);
         // create owner's signature
@@ -168,16 +156,10 @@ contract ERC20UbiquityTest is LocalTestHelper {
         );
         (uint8 v, bytes32 r, bytes32 s) = vm.sign(ownerPrivateKey, digest);
         // run permit
-        uint noncesBefore = ERC20Ubiquity(token_addr).nonces(owner);
+        uint256 noncesBefore = ERC20Ubiquity(token_addr).nonces(owner);
         vm.prank(spender);
         ERC20Ubiquity(token_addr).permit(
-            owner,
-            spender,
-            1e18,
-            block.timestamp + 1 days,
-            v,
-            r,
-            s
+            owner, spender, 1e18, block.timestamp + 1 days, v, r, s
         );
         assertEq(ERC20Ubiquity(token_addr).allowance(owner, spender), 1e18);
         assertEq(ERC20Ubiquity(token_addr).nonces(owner), noncesBefore + 1);
@@ -268,7 +250,7 @@ contract ERC20UbiquityTest is LocalTestHelper {
 
     function testMint_ShouldMintTokens() public {
         address mockAddress = address(0x1);
-        uint balanceBefore = ERC20Ubiquity(token_addr).balanceOf(mockAddress);
+        uint256 balanceBefore = ERC20Ubiquity(token_addr).balanceOf(mockAddress);
         vm.prank(admin);
         vm.expectEmit(true, true, false, true);
         emit Minting(mockAddress, admin, 100);
@@ -314,7 +296,8 @@ contract ERC20UbiquityTest is LocalTestHelper {
     function testTransfer_ShouldRevert_IfContractIsPaused() public {
         // deploy contract with exposed internal methods
         vm.prank(admin);
-        ERC20UbiquityHarness erc20Ubiquity = new ERC20UbiquityHarness(uad_manager_addr, "Test", "Test");
+        ERC20UbiquityHarness erc20Ubiquity =
+            new ERC20UbiquityHarness(uad_manager_addr, "Test", "Test");
         // admin pauses contract
         vm.prank(admin);
         erc20Ubiquity.pause();
@@ -328,7 +311,8 @@ contract ERC20UbiquityTest is LocalTestHelper {
     function testTransfer_ShouldTransferTokens() public {
         // deploy contract with exposed internal methods
         vm.prank(admin);
-        ERC20UbiquityHarness erc20Ubiquity = new ERC20UbiquityHarness(uad_manager_addr, "Test", "Test");
+        ERC20UbiquityHarness erc20Ubiquity =
+            new ERC20UbiquityHarness(uad_manager_addr, "Test", "Test");
         // mint tokens to admin
         vm.prank(admin);
         erc20Ubiquity.mint(admin, 100);
